add sorting for projects by name and refactor code
Feature #14422_item_sorting_in_copy_to_project_modal

Arvados-DCO-1.1-Signed-off-by: Janicki Artur <artur.janicki@contractors.roche.com>
